Request validation moved to controller
Service top level removed since redudant with controller.
Landing page HTML moved to separate template file.
Dead code removed.
