+++ backend/tests/views/test_api_routes.py
@@ -51,64 +51,11 @@ class TestAPIEndpoints:
         assert "endpoints" in data
         assert isinstance(data["endpoints"], list)
 
-    @patch("backend.views.api_routes.SUMMARIZATION_SERVICE")
-    def test_intent_endpoint_success(self, mock_service, client):
-        """Test successful intent prediction"""
-        # Mock service response
-        mock_service.classify_intent.return_value = {
-            "intent": "Technology",
-            "confidence": 0.85,
-            "timestamp": "2024-01-01T00:00:00Z"
-        }
-
-        response = client.post(
-            "/intent",
-            json={"text": "Tell me about artificial intelligence"},
-            content_type="application/json",
-        )
-
-        assert response.status_code == 200
-        data = json.loads(response.data)
-        assert data["intent"] == "Technology"
-        assert data["confidence"] == 0.85
-        assert data["model_type"] == "TensorFlow LSTM"
-        assert "timestamp" in data
-
-    def test_intent_endpoint_missing_text(self, client):
-        """Test intent endpoint with missing text parameter"""
-        response = client.post("/intent", json={}, content_type="application/json")
-
-        assert response.status_code == 400
-        data = json.loads(response.data)
-        assert "error" in data
-        assert 'Missing "text" field' in data["error"]
-
-    def test_intent_endpoint_empty_text(self, client):
-        """Test intent endpoint with empty text"""
-        response = client.post(
-            "/intent", json={"text": ""}, content_type="application/json"
-        )
-
-        assert response.status_code == 400
-        data = json.loads(response.data)
-        assert "error" in data
-        assert "Text field cannot be empty" in data["error"]
-
-    def test_intent_endpoint_invalid_json(self, client):
-        """Test intent endpoint with invalid JSON"""
-        # Send malformed JSON
-        response = client.post(
-            "/intent", data="invalid json", content_type="application/json"
-        )
-
-        # Should handle the error gracefully
-        assert response.status_code in [400, 500]  # Accept either error code
-
-    @patch("backend.views.api_routes.SUMMARIZATION_SERVICE")
-    def test_intent_bert_endpoint_success(self, mock_service, client):
+    @patch("backend.views.api_routes.SUMMARIZATION_CONTROLLER")
+    def test_intent_bert_endpoint_success(self, mock_controller, client):
         """Test successful BERT intent prediction"""
         # Mock service response
-        mock_service.classify_intent.return_value = {
+        mock_controller.classify_intent.return_value = {
             "intent": "Science",
             "confidence": 0.92,
             "model_type": "BERT",
@@ -127,11 +74,11 @@ class TestAPIEndpoints:
         assert data["confidence"] == 0.92
         assert data["model_type"] == "BERT"
 
-    @patch("backend.views.api_routes.SUMMARIZATION_SERVICE")
-    def test_intent_bert_endpoint_model_not_loaded(self, mock_service, client):
+    @patch("backend.views.api_routes.SUMMARIZATION_CONTROLLER")
+    def test_intent_bert_endpoint_model_not_loaded(self, mock_controller, client):
         """Test BERT intent endpoint when model is not loaded"""
         # Mock service response with error
-        mock_service.classify_intent.return_value = {
+        mock_controller.classify_intent.return_value = {
             "error": "BERT model not loaded"
         }
 
@@ -145,11 +92,11 @@ class TestAPIEndpoints:
         data = json.loads(response.data)
         assert "error" in data
 
-    @patch("backend.views.api_routes.SUMMARIZATION_SERVICE")
-    def test_summarize_endpoint_success(self, mock_service, client):
+    @patch("backend.views.api_routes.SUMMARIZATION_CONTROLLER")
+    def test_summarize_endpoint_success(self, mock_controller, client):
         """Test successful single-source summarization"""
         # Mock service response
-        mock_service.summarize_single_source.return_value = {
+        mock_controller.summarize_single_source.return_value = {
             "query": "Apollo 11",
             "summary": "Apollo 11 was the first manned mission to land on the Moon.",
             "intent": {"category": "History", "confidence": 0.85},
@@ -191,11 +138,11 @@ class TestAPIEndpoints:
         assert "error" in data
         assert "Empty query provided" in data["error"]
 
-    @patch("backend.views.api_routes.SUMMARIZATION_SERVICE")
-    def test_summarize_multi_source_success(self, mock_service, client):
+    @patch("backend.views.api_routes.SUMMARIZATION_CONTROLLER")
+    def test_summarize_multi_source_success(self, mock_controller, client):
         """Test successful multi-source summarization"""
         # Mock service response
-        mock_service.summarize_multi_source_with_agents.return_value = {
+        mock_controller.summarize_multi_source_with_agents.return_value = {
             "query": "Space exploration",
             "summary": "Space exploration has been a major human endeavor.",
             "intent": "Science",
@@ -227,11 +174,11 @@ class TestAPIEndpoints:
         assert data["intent"] == "Science"
         assert len(data["articles"]) == 3
 
-    @patch("backend.views.api_routes.SUMMARIZATION_SERVICE")
-    def test_summarize_multi_source_error(self, mock_service, client):
+    @patch("backend.views.api_routes.SUMMARIZATION_CONTROLLER")
+    def test_summarize_multi_source_error(self, mock_controller, client):
         """Test multi-source summarization with error"""
         # Mock service response with error
-        mock_service.summarize_multi_source_with_agents.return_value = {
+        mock_controller.summarize_multi_source_with_agents.return_value = {
             "error": "No articles found for query"
         }
 
@@ -256,24 +203,10 @@ class TestAPIErrorHandling:
         with app.test_client() as client:
             yield client
 
-    @patch("backend.views.api_routes.SUMMARIZATION_SERVICE")
-    def test_intent_endpoint_exception(self, mock_service, client):
-        """Test intent endpoint when classifier raises exception"""
-        mock_service.classify_intent.side_effect = Exception("Service error")
-
-        response = client.post(
-            "/intent", json={"text": "test text"}, content_type="application/json"
-        )
-
-        assert response.status_code == 500
-        data = json.loads(response.data)
-        assert "error" in data
-        assert "Internal server error" in data["error"]
-
-    @patch("backend.views.api_routes.SUMMARIZATION_SERVICE")
-    def test_intent_bert_endpoint_exception(self, mock_service, client):
+    @patch("backend.views.api_routes.SUMMARIZATION_CONTROLLER")
+    def test_intent_bert_endpoint_exception(self, mock_controller, client):
         """Test BERT intent endpoint when classifier raises exception"""
-        mock_service.classify_intent.side_effect = Exception("Service error")
+        mock_controller.classify_intent.side_effect = Exception("Service error")
 
         response = client.post(
             "/intent_bert", json={"text": "test text"}, content_type="application/json"
@@ -284,10 +217,10 @@ class TestAPIErrorHandling:
         assert "error" in data
         assert "Internal server error" in data["error"]
 
-    @patch("backend.views.api_routes.SUMMARIZATION_SERVICE")
-    def test_summarize_endpoint_exception(self, mock_service, client):
+    @patch("backend.views.api_routes.SUMMARIZATION_CONTROLLER")
+    def test_summarize_endpoint_exception(self, mock_controller, client):
         """Test summarize endpoint when summarizer raises exception"""
-        mock_service.summarize_single_source.side_effect = Exception("Service error")
+        mock_controller.summarize_single_source.side_effect = Exception("Service error")
 
         response = client.post(
             "/summarize", json={"query": "test query"}, content_type="application/json"
@@ -308,50 +241,10 @@ class TestAPIResponseFormat:
         with app.test_client() as client:
             yield client
 
-    @patch("backend.views.api_routes.SUMMARIZATION_SERVICE")
-    def test_intent_response_format(self, mock_service, client):
-        """Test that intent response has correct format"""
-        mock_service.classify_intent.return_value = {
-            "intent": "Technology",
-            "confidence": 0.85,
-            "model_type": "TensorFlow LSTM",
-            "model_loaded": True,
-            "categories_available": ["Technology", "Science", "History"],
-            "timestamp": "2024-01-01T00:00:00Z"
-        }
-
-        response = client.post(
-            "/intent", json={"text": "AI technology"}, content_type="application/json"
-        )
-
-        data = json.loads(response.data)
-
-        # Check required fields
-        required_fields = [
-            "text",
-            "intent",
-            "confidence",
-            "model_type",
-            "model_loaded",
-            "categories_available",
-            "timestamp",
-        ]
-        for field in required_fields:
-            assert field in data
-
-        # Check data types
-        assert isinstance(data["text"], str)
-        assert isinstance(data["intent"], str)
-        assert isinstance(data["confidence"], (int, float))
-        assert isinstance(data["model_type"], str)
-        assert isinstance(data["model_loaded"], bool)
-        assert isinstance(data["categories_available"], list)
-        assert isinstance(data["timestamp"], str)
-
-    @patch("backend.views.api_routes.SUMMARIZATION_SERVICE")
-    def test_intent_bert_response_format(self, mock_service, client):
+    @patch("backend.views.api_routes.SUMMARIZATION_CONTROLLER")
+    def test_intent_bert_response_format(self, mock_controller, client):
         """Test that BERT intent response has correct format"""
-        mock_service.classify_intent.return_value = {
+        mock_controller.classify_intent.return_value = {
             "intent": "Science",
             "confidence": 0.92,
             "model_type": "BERT",
@@ -415,36 +308,10 @@ class TestAPIIntegration:
             assert response.status_code == 200
             assert "application/json" in response.content_type
 
-    @patch("backend.views.api_routes.SUMMARIZATION_SERVICE")
-    def test_confidence_score_bounds(self, mock_service, client):
-        """Test that confidence scores are within valid bounds"""
-        # Test various confidence values
-        test_cases = [
-            ("Technology", 0.0),
-            ("Science", 0.5),
-            ("History", 1.0),
-            ("General", 0.999),
-        ]
-
-        for intent, confidence in test_cases:
-            mock_service.classify_intent.return_value = {
-                "intent": intent,
-                "confidence": confidence,
-                "timestamp": "2024-01-01T00:00:00Z"
-            }
-
-            response = client.post(
-                "/intent", json={"text": "test text"}, content_type="application/json"
-            )
-
-            data = json.loads(response.data)
-            assert 0.0 <= data["confidence"] <= 1.0
-
     def test_error_response_format(self, client):
         """Test that error responses have consistent format"""
         # Test various error conditions
         error_cases = [
-            ("/intent", {}),  # Missing text
             ("/intent_bert", {}),  # Missing text
             ("/summarize", {}),  # Missing query
             ("/summarize_multi_source", {}),  # Missing query
